requirements.txt
#kaggle
mlflow
dvc
dagshub
dvc-s3
pandas
scikit-learn
numpy
python-dotenv
datetime
seaborn
matplotlib
ipykernel
ruamel.yaml
evidently
flask
#huggingface_hub 
google-cloud-storage